update build system
I wanted something that is cross-platform out of the box.
Make is itself cross platform, but as it relies on the underlying shell, it is not cross platform in practice.

I figured most GBdev'ers have Python installed already (I've seen some of the scripts people have lying around in their repos). So I rewrote the functionality of the old Makefile in Python, and integrated it into the build step.